feat: Implemented assembly def, initial clean up file structure
- Also added test & editor assembly def, rough organize namespace
