--- Assets/Scripts/ApplicationSettings.cs
@@ -1,12 +1,0 @@
-using UnityEngine;
-
-public class ApplicationSettings : MonoBehaviour
-{
-
-    void Start()
-    {
-        QualitySettings.vSyncCount = 0; // Disable VSync
-        Application.targetFrameRate = 60;
-    }
-
-}
