+++ Assets/Scripts/Editor/RidersEditor.asmdef
@@ -1,0 +1,22 @@
+{
+    "name": "RidersEditor",
+    "rootNamespace": "RidersEditor",
+    "references": [
+        "GUID:343deaaf83e0cee4ca978e7df0b80d21",
+        "GUID:531d3150222f5b049b95855bf8783e30",
+        "GUID:f7200fc466f3ff943a2bfd1e542f785c"
+    ],
+    "includePlatforms": [
+        "Editor",
+        "WindowsStandalone32",
+        "WindowsStandalone64"
+    ],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
+}
