+++ Assets/Scripts/Editor/RidersEditor.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: d0b18f2bc867fd2429371388d4f1c18a
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
