--- Assets/Scripts/Enums.cs
@@ -1,9 +1,0 @@
-public static class Enums
-{
-    public enum VehicleType
-    {
-        Skateboard,
-        Blades,
-        Bike
-    }
-}
