--- Assets/Scripts/RacerInformation.cs
@@ -1,11 +1,0 @@
-using UnityEngine;
-using UnityEngine.TextCore.Text;
-
-// questionable, find examples elsewhere or run by James
-[System.Serializable]
-public struct RacerInformation
-{
-    public BaseVehicle VehicleSelected;
-    public int CharacterSelected;
-    public bool IsPlayer;
-}
