+++ Assets/Scripts/RidersGame.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bff3b187fd0a46e458bb66057b7f11ac
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
