+++ Assets/Scripts/RidersGame/Analytics.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 39b9555a169df954cb4a0637c4060da4
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
