+++ Assets/Scripts/RidersGame/Enums.cs
@@ -1,0 +1,9 @@
+namespace RidersCore.Data
+{
+    public enum VehicleType
+        {
+            Skateboard,
+            Blades,
+            Bike
+        }
+}
