+++ Assets/Scripts/RidersGame/Inputs.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b4d339ffecba80b4caa02fed6df6fe0c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
