+++ Assets/Scripts/RidersGame/Inputs/InputFrameRecord.cs
@@ -1,0 +1,15 @@
+namespace RidersCore
+{
+    [System.Serializable]
+    public struct InputFrameRecord
+    {
+        public int Frame;
+        public ActorInputData Input;
+
+        public InputFrameRecord(int frame, ActorInputData input)
+        {
+            Frame = frame;
+            Input = input;
+        }
+    }
+}
