+++ Assets/Scripts/RidersGame/RidersCore.Unity.asmdef
@@ -1,0 +1,20 @@
+{
+    "name": "RidersRuntime",
+    "rootNamespace": "RidersCore",
+    "references": [
+        "GUID:75469ad4d38634e559750d17036d5f7c",
+        "GUID:21d1eb854b91ade49bc69a263d12bee2",
+        "GUID:d8b63aba1907145bea998dd612889d6b",
+        "GUID:2bafac87e7f4b9b418d9448d219b01ab",
+        "GUID:6055be8ebefd69e48b49212b09b47b2f"
+    ],
+    "includePlatforms": [],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": [],
+    "versionDefines": [],
+    "noEngineReferences": false
+}
