+++ Assets/Scripts/RidersGame/RidersCore.Unity.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: f7200fc466f3ff943a2bfd1e542f785c
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
