+++ Assets/Scripts/RidersGame/System.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ff1805e1985741c4d9371d1cf6221a21
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
