+++ Assets/Scripts/RidersGame/System/ApplicationSettings.cs
@@ -1,0 +1,15 @@
+using UnityEngine;
+
+namespace RidersCore
+{
+    public class ApplicationSettings : MonoBehaviour
+    {
+
+        void Start()
+        {
+            QualitySettings.vSyncCount = 0; // Disable VSync
+            Application.targetFrameRate = 60;
+        }
+
+    }
+}
