+++ Assets/Scripts/RidersGame/System/ApplicationSettings.cs.meta
