+++ Assets/Scripts/RidersGame/VehicleSystems/BikeVehicle.cs
@@ -1,13 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using UnityEngine;
 
-namespace Assets.Scripts.VehicleSystems
+namespace RidersCore
 {
-    class BoardVehicle : BaseVehicle
+    class BikeVehicle : BaseVehicle
     {
         //// replace base stats
         //public new BaseVehicle.Stats baseStats = new BaseVehicle.Stats
@@ -26,15 +21,12 @@ namespace Assets.Scripts.VehicleSystems
         //    BoostAccel = 15f,
         //};
 
-        [Header("Skateboard Specific Settings")]
+        [Header("Bike Specific Settings")]
         // Nothing for now
 
         [Header("Physical Wheels")]
-        [Tooltip("The physical representations of the skateboard's wheels.")]
-        public WheelCollider FrontLeftWheel;
-        public WheelCollider FrontRightWheel;
-        public WheelCollider BackLeftWheel;
-        public WheelCollider BackRightWheel;
-
+        [Tooltip("The physical representations of the bike's wheels.")]
+        public WheelCollider FrontWheel;
+        public WheelCollider BackWheel;
     }
 }
