+++ Assets/Scripts/RidersGame/VehicleSystems/BladesVehicle.cs
@@ -1,11 +1,6 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using UnityEngine;
 
-namespace Assets.Scripts.VehicleSystems
+namespace RidersCore
 {
     class BladesVehicle : BaseVehicle
     {
