+++ Assets/Scripts/RidersGame/VehicleSystems/BoardVehicle.cs
@@ -1,13 +1,8 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using UnityEngine;
+﻿using UnityEngine;
 
-namespace Assets.Scripts.VehicleSystems
+namespace RidersCore
 {
-    class BikeVehicle : BaseVehicle
+    class BoardVehicle : BaseVehicle
     {
         //// replace base stats
         //public new BaseVehicle.Stats baseStats = new BaseVehicle.Stats
@@ -26,12 +21,15 @@ namespace Assets.Scripts.VehicleSystems
         //    BoostAccel = 15f,
         //};
 
-        [Header("Bike Specific Settings")]
+        [Header("Skateboard Specific Settings")]
         // Nothing for now
 
         [Header("Physical Wheels")]
-        [Tooltip("The physical representations of the bike's wheels.")]
-        public WheelCollider FrontWheel;
-        public WheelCollider BackWheel;
+        [Tooltip("The physical representations of the skateboard's wheels.")]
+        public WheelCollider FrontLeftWheel;
+        public WheelCollider FrontRightWheel;
+        public WheelCollider BackLeftWheel;
+        public WheelCollider BackRightWheel;
+
     }
 }
