--- Assets/Scripts/VehicleSystems/Inputs/InputFrameRecord.cs
@@ -1,12 +1,0 @@
-[System.Serializable]
-public struct InputFrameRecord
-{
-    public int Frame;
-    public ActorInputData Input;
-
-    public InputFrameRecord(int frame, ActorInputData input)
-    {
-        Frame = frame;
-        Input = input;
-    }
-}
